requirements.txt
streamlit
python-docx
pdfplumber
pandas